Move uint_from_xof to `uint::traits`
